aggiunti controlli nella creazione dei prodotti

+++ shop/Videogame.java
@@ -4,16 +4,16 @@ public class Videogame
 {
     private String title;
     private double price;
-    private int quantity;
+    private double quantity;
     private static int previousid = -1;
     private int id;
 
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 
-    public Videogame(String title, double price, int quantity) {
+    public Videogame(String title, double price, double quantity) {
         this.title = title;
         this.quantity = quantity;
         this.price = price;
@@ -25,7 +25,7 @@ public class Videogame
         return title;
     }
 
-    public int getQuantity(){ return quantity; }
+    public double getQuantity(){ return quantity; }
 
     public double getPrice()
